+++ pkg/controllers/controllers_test.go
@@ -1,4 +1,4 @@
-package handlers
+package controllers
 
 import (
 	"bytes"
